division literal

+++ src/factors/factor.cpp
@@ -4,19 +4,19 @@ Factor* Factor::Create(const std::vector<Token> tokens, const FactorType& type)
 {
 	switch (type)
 	{
-    case (int)Factor::FactorType::Literal:
+    case Factor::FactorType::Literal:
 		return new Literal(tokens);
 		break;
-    case (int)Factor::FactorType::Exponential:
+    case Factor::FactorType::Exponential:
 		return new Exponential(tokens);
 		break;
-    case (int)Factor::FactorType::Fraction:
+    case Factor::FactorType::Fraction:
 		return new Fraction(tokens);
 		break;
-    case (int)Factor::FactorType::Parenthesis:
+    case Factor::FactorType::Parenthesis:
 		return new Parenthesis(tokens);
 		break;
-    case (int)Factor::FactorType::Invalid:
+    case Factor::FactorType::Invalid:
 		return nullptr;
 		break;
 	default:
